Support queries with exact matches
Algorithm:
Run a query as normal, but also check if an exact match occurred.
If one or more exact matches occurred, then there may be extra
nodes that match to a list but not part of the exact match. Traverse
the result tree and mark the nodes that are the exact match and their
parent nodes to the top of the result tree. Copy the marked nodes to
a new tree and delete the old result tree.